+++ classmap.php
@@ -66,6 +66,7 @@ require_once ALGOLIA_PATH . 'includes/utilities/class-algolia-update-messages.ph
 
 if ( is_admin() ) {
 	require_once ALGOLIA_PATH . 'includes/admin/class-algolia-admin.php';
+	require_once ALGOLIA_PATH . 'includes/admin/class-algolia-admin-field-helpers.php';
 	require_once ALGOLIA_PATH . 'includes/admin/class-algolia-admin-page-settings.php';
 	require_once ALGOLIA_PATH . 'includes/admin/class-algolia-admin-page-autocomplete.php';
 	require_once ALGOLIA_PATH . 'includes/admin/class-algolia-admin-page-native-search.php';